Allowing for revalidation of all vs single pages
