+++ package.json
@@ -5,7 +5,6 @@
         "dev": "next dev",
         "build": "next build",
         "build:analyze": "ANALYZE=true next build",
-        "crawl": "node src/scripts/crawl-devcenter.js",
         "start": "next start",
         "lint": "next lint",
         "format:write": "prettier --write .",
@@ -36,7 +35,7 @@
         "html-react-parser": "1.4.11",
         "lodash.debounce": "4.0.8",
         "lru-cache": "7.10.1",
-        "next": "12.2.0",
+        "next": "12.1.0",
         "next-plugin-preval": "1.2.4",
         "next-seo": "5.4.0",
         "pino": "8.0.0",
